# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AquaQuest API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type QuestDifficulty =
  (typeof QuestDifficulty)[keyof typeof QuestDifficulty];

export const QuestDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
  legendary: "legendary",
} as const;

export type QuestStatus = (typeof QuestStatus)[keyof typeof QuestStatus];

export const QuestStatus = {
  active: "active",
  completed: "completed",
  failed: "failed",
} as const;

export interface Quest {
  id: number;
  title: string;
  description: string;
  difficulty: QuestDifficulty;
  status: QuestStatus;
  xpReward: number;
  depthLevel: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateQuestInputDifficulty =
  (typeof CreateQuestInputDifficulty)[keyof typeof CreateQuestInputDifficulty];

export const CreateQuestInputDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
  legendary: "legendary",
} as const;

export interface CreateQuestInput {
  title: string;
  description: string;
  difficulty: CreateQuestInputDifficulty;
  xpReward: number;
  depthLevel: number;
}

export type UpdateQuestInputDifficulty =
  (typeof UpdateQuestInputDifficulty)[keyof typeof UpdateQuestInputDifficulty];

export const UpdateQuestInputDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
  legendary: "legendary",
} as const;

export type UpdateQuestInputStatus =
  (typeof UpdateQuestInputStatus)[keyof typeof UpdateQuestInputStatus];

export const UpdateQuestInputStatus = {
  active: "active",
  completed: "completed",
  failed: "failed",
} as const;

export interface UpdateQuestInput {
  title?: string;
  description?: string;
  difficulty?: UpdateQuestInputDifficulty;
  status?: UpdateQuestInputStatus;
  xpReward?: number;
  depthLevel?: number;
}

export type CreatureRarity =
  (typeof CreatureRarity)[keyof typeof CreatureRarity];

export const CreatureRarity = {
  common: "common",
  uncommon: "uncommon",
  rare: "rare",
  epic: "epic",
  legendary: "legendary",
} as const;

export interface Creature {
  id: number;
  name: string;
  species: string;
  rarity: CreatureRarity;
  depthFound: number;
  discoveredAt: string;
  description: string;
  emoji: string;
}

export type AchievementCategory =
  (typeof AchievementCategory)[keyof typeof AchievementCategory];

export const AchievementCategory = {
  exploration: "exploration",
  combat: "combat",
  collection: "collection",
  social: "social",
} as const;

export interface Achievement {
  id: number;
  title: string;
  description: string;
  icon: string;
  unlockedAt?: string;
  category: AchievementCategory;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type UserProfileExperienceLevel =
  (typeof UserProfileExperienceLevel)[keyof typeof UserProfileExperienceLevel];

export const UserProfileExperienceLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface UserProfile {
  id: number;
  age: number;
  height: string;
  weight: string;
  goal: string;
  experienceLevel: UserProfileExperienceLevel;
  createdAt: string;
  updatedAt: string;
}

export type UserProfileInputExperienceLevel =
  (typeof UserProfileInputExperienceLevel)[keyof typeof UserProfileInputExperienceLevel];

export const UserProfileInputExperienceLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface UserProfileInput {
  age: number;
  height: string;
  weight: string;
  goal: string;
  experienceLevel: UserProfileInputExperienceLevel;
}

export interface Exercise {
  name: string;
  sets: number;
  reps: string;
  rest: string;
  notes?: string;
}

export interface WorkoutDay {
  day: string;
  focus: string;
  exercises: Exercise[];
}

export interface WorkoutPlan {
  id: number;
  profileId?: number;
  plan: WorkoutDay[];
  createdAt: string;
}

export interface ExerciseLogInput {
  workoutPlanId: number;
  dayIndex: number;
  exerciseName: string;
  completed: boolean;
}

export interface ExerciseLog {
  id: number;
  workoutPlanId: number;
  dayIndex: number;
  exerciseName: string;
  completed: boolean;
  completedAt?: string;
}