# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AquaQuest API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: quests
    description: Quest operations
  - name: creatures
    description: Sea creature operations
  - name: achievements
    description: Achievement operations
  - name: workouts
    description: Workout plan operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /quests:
    get:
      operationId: getQuests
      tags: [quests]
      summary: Get all quests
      responses:
        "200":
          description: List of quests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quest"
    post:
      operationId: createQuest
      tags: [quests]
      summary: Create a new quest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestInput"
      responses:
        "201":
          description: Quest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
  /quests/{id}:
    get:
      operationId: getQuest
      tags: [quests]
      summary: Get a quest by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quest found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
        "404":
          description: Quest not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateQuest
      tags: [quests]
      summary: Update a quest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestInput"
      responses:
        "200":
          description: Quest updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
    delete:
      operationId: deleteQuest
      tags: [quests]
      summary: Delete a quest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Quest deleted
  /creatures:
    get:
      operationId: getCreatures
      tags: [creatures]
      summary: Get all discovered sea creatures
      responses:
        "200":
          description: List of sea creatures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Creature"
  /achievements:
    get:
      operationId: getAchievements
      tags: [achievements]
      summary: Get all achievements
      responses:
        "200":
          description: List of achievements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Achievement"
  /workouts/profile:
    get:
      operationId: getWorkoutProfile
      tags: [workouts]
      summary: Get user workout profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: saveWorkoutProfile
      tags: [workouts]
      summary: Save or update user workout profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /workouts/generate:
    post:
      operationId: generateWorkoutPlan
      tags: [workouts]
      summary: Generate AI-powered 7-day workout plan
      responses:
        "200":
          description: Generated workout plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlan"
        "400":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workouts/plans:
    get:
      operationId: getWorkoutPlans
      tags: [workouts]
      summary: Get all saved workout plans
      responses:
        "200":
          description: List of workout plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutPlan"
  /workouts/plans/{id}:
    get:
      operationId: getWorkoutPlan
      tags: [workouts]
      summary: Get a specific workout plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workout plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlan"
        "404":
          description: Plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workouts/logs:
    post:
      operationId: logExercise
      tags: [workouts]
      summary: Log a completed exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseLogInput"
      responses:
        "200":
          description: Exercise logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseLog"
  /workouts/logs/{planId}:
    get:
      operationId: getWorkoutLogs
      tags: [workouts]
      summary: Get exercise logs for a plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExerciseLog"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Quest:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        status:
          type: string
          enum: [active, completed, failed]
        xpReward:
          type: integer
        depthLevel:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - difficulty
        - status
        - xpReward
        - depthLevel
        - createdAt
        - updatedAt
    CreateQuestInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        xpReward:
          type: integer
        depthLevel:
          type: integer
      required:
        - title
        - description
        - difficulty
        - xpReward
        - depthLevel
    UpdateQuestInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard, legendary]
        status:
          type: string
          enum: [active, completed, failed]
        xpReward:
          type: integer
        depthLevel:
          type: integer
    Creature:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        species:
          type: string
        rarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
        depthFound:
          type: integer
        discoveredAt:
          type: string
          format: date-time
        description:
          type: string
        emoji:
          type: string
      required:
        - id
        - name
        - species
        - rarity
        - depthFound
        - discoveredAt
        - description
        - emoji
    Achievement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        unlockedAt:
          type: string
          format: date-time
        category:
          type: string
          enum: [exploration, combat, collection, social]
      required:
        - id
        - title
        - description
        - icon
        - category
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        age:
          type: integer
        height:
          type: string
        weight:
          type: string
        goal:
          type: string
        experienceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - age
        - height
        - weight
        - goal
        - experienceLevel
        - createdAt
        - updatedAt
    UserProfileInput:
      type: object
      properties:
        age:
          type: integer
        height:
          type: string
        weight:
          type: string
        goal:
          type: string
        experienceLevel:
          type: string
          enum: [beginner, intermediate, advanced]
      required:
        - age
        - height
        - weight
        - goal
        - experienceLevel
    Exercise:
      type: object
      properties:
        name:
          type: string
        sets:
          type: integer
        reps:
          type: string
        rest:
          type: string
        notes:
          type: string
      required:
        - name
        - sets
        - reps
        - rest
    WorkoutDay:
      type: object
      properties:
        day:
          type: string
        focus:
          type: string
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/Exercise"
      required:
        - day
        - focus
        - exercises
    WorkoutPlan:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        plan:
          type: array
          items:
            $ref: "#/components/schemas/WorkoutDay"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - plan
        - createdAt
    ExerciseLogInput:
      type: object
      properties:
        workoutPlanId:
          type: integer
        dayIndex:
          type: integer
        exerciseName:
          type: string
        completed:
          type: boolean
      required:
        - workoutPlanId
        - dayIndex
        - exerciseName
        - completed
    ExerciseLog:
      type: object
      properties:
        id:
          type: integer
        workoutPlanId:
          type: integer
        dayIndex:
          type: integer
        exerciseName:
          type: string
        completed:
          type: boolean
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - workoutPlanId
        - dayIndex
        - exerciseName
        - completed