# lib/api-zod/src/generated/types/exerciseLogInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AquaQuest API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExerciseLogInput {
  workoutPlanId: number;
  dayIndex: number;
  exerciseName: string;
  completed: boolean;
}