# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AquaQuest API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileExperienceLevel } from "./userProfileExperienceLevel";

export interface UserProfile {
  id: number;
  age: number;
  height: string;
  weight: string;
  goal: string;
  experienceLevel: UserProfileExperienceLevel;
  createdAt: Date;
  updatedAt: Date;
}