# lib/api-zod/src/generated/types/userProfileInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AquaQuest API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileInputExperienceLevel } from "./userProfileInputExperienceLevel";

export interface UserProfileInput {
  age: number;
  height: string;
  weight: string;
  goal: string;
  experienceLevel: UserProfileInputExperienceLevel;
}