# lib/api-zod/src/generated/types/workoutDay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AquaQuest API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";

export interface WorkoutDay {
  day: string;
  focus: string;
  exercises: Exercise[];
}