# lib/api-zod/src/generated/types/workoutPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AquaQuest API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutDay } from "./workoutDay";

export interface WorkoutPlan {
  id: number;
  profileId?: number;
  plan: WorkoutDay[];
  createdAt: Date;
}